Better hit box
- Better melee hit box which spawn in the direction the player/mob is facing
- Mob can now damage the player once every second for 1 dmg
- Items now need to be picked up using space

+++ interactions.py
@@ -1,33 +1,51 @@
 from helper import *
+from tools import *
 
 game_folder = path.dirname(__file__)
 music_folder = path.join(game_folder, 'resources')
 
+class MeleeHitBox():
+    def __init__(self, game, x, y, width=0, height=0):
+        self.game = game
+        self.hitbox = pg.Rect(x + width, y + height, 64, 64)
+
+    def draw_hitbox(self, game, color):
+        pg.draw.rect(game.screen, color, self.hitbox)
+
 class Attack():
     def __init__(self, game, target, weapon):
-        targets = pg.sprite.groupcollide(target, weapon, False, False, collided)
-        now = pg.time.get_ticks()
-        for target in targets:
-            if now - game.last_hit > TOOL_LIFETIME:
-                game.last_hit = now
-                target.attacked = True
-                target.health -= 25
-                # print(type(target).__name__)
+        sprites = target.sprites()
+        self.game = game
+        self.targets = []
+        for sprite in sprites:
+            if collided(weapon, sprite):
+                self.targets.append(sprite)
+        self.now = pg.time.get_ticks()
+
+    def target_hit(self):
+        for target in self.targets:
+            if self.now - self.game.last_hit > TOOL_LIFETIME:
+                self.now = pg.time.get_ticks()
+                self.game.last_hit = self.now
                 if type(target).__name__ == 'Tree' and not target.chopped:
                     index = random.randint(0, len(WOOD_CHOP) - 1)
                     effect = pg.mixer.Sound(path.join(music_folder, WOOD_CHOP[index]))
                     effect.play()
+                if hasattr(target, 'attacked'):
+                    target.attacked = True
+                return target
 
 class Pickup():
-    def __init__(self, game, target):
-        pickups = pg.sprite.spritecollide(game.player, target, False, collided)
+    def __init__(self, game, target, hitbox):
+        pickups = target.sprites()
         for pickup in pickups:
-            if not pickup.in_inventory:
-                index = game.inventory.add(pickup.item_name)
-                game.hotbar.add_to_hotbar(index, pickup.item_name)
-                if pickup.item_name == 'rock': # temp rock spawn before cave level
-                    game.rock_count -= 1
-                    index = random.randint(0, len(ROCK_MINE) - 1)
-                    effect = pg.mixer.Sound(path.join(music_folder, ROCK_MINE[index]))
-                    effect.play()
-                pickup.kill()
+            if collided(hitbox, pickup):
+                if not pickup.in_inventory:
+                    index = game.inventory.add(pickup.item_name)
+                    game.hotbar.add_to_hotbar(index, pickup.item_name)
+                    if pickup.item_name == 'rock': # temp rock spawn before cave level
+                        game.rock_count -= 1
+                        index = random.randint(0, len(ROCK_MINE) - 1)
+                        effect = pg.mixer.Sound(path.join(music_folder, ROCK_MINE[index]))
+                        effect.play()
+                    pickup.kill()
